add root login page, dashboard path, refactor middleware to only target dash
