+++ app/(kiosk dashboard)/dashboard/@sidebar/default.tsx
