+++ app/(kiosk dashboard)/dashboard/@sidebar/issues/[id]/page.tsx
