+++ app/(kiosk dashboard)/dashboard/@sidebar/issues/page.tsx
