+++ app/(kiosk dashboard)/dashboard/@sidebar/kioskHealthUpdater.tsx
@@ -2,9 +2,9 @@
 
 import { useEffect } from 'react';
 import { useSetRecoilState } from 'recoil';
-import { getHealthStatuses } from '../../../helpers/httpMethods';
-import throwError from '../../../helpers/throwError';
-import { kioskHealthStatusesState } from '../../../state/sidebarState';
+import { getHealthStatuses } from '../../../../helpers/httpMethods';
+import throwError from '../../../../helpers/throwError';
+import { kioskHealthStatusesState } from '../../../../state/sidebarState';
 
 const HEALTH_FETCH_INTERVAL = parseInt(process.env.NEXT_PUBLIC_HEALTH_FETCH_INTERVAL ?? throwError('NEXT_PUBLIC_HEALTH_FETCH_INTERVAL is not defined'), 10);
