+++ app/(kiosk dashboard)/dashboard/@sidebar/kioskMap.module.css
