+++ app/(kiosk dashboard)/dashboard/@sidebar/led/page.tsx
