+++ app/(kiosk dashboard)/dashboard/@sidebar/page.tsx
