+++ app/(kiosk dashboard)/dashboard/@sidebar/sidebarRoot.tsx
@@ -2,9 +2,9 @@
 
 import { ReactNode } from 'react';
 import { MutableSnapshot, RecoilRoot } from 'recoil';
-import { Kiosk } from '../../../sanity.types';
-import { allKiosksState, kioskHealthStatusesState } from '../../../state/sidebarState';
-import { ServerHealthStatuses } from '../../../types/serverHealthStatuses';
+import { Kiosk } from '../../../../sanity.types';
+import { allKiosksState, kioskHealthStatusesState } from '../../../../state/sidebarState';
+import { ServerHealthStatuses } from '../../../../types/serverHealthStatuses';
 import KioskHealthUpdater from './kioskHealthUpdater';
 
 interface Props {
