+++ app/(kiosk dashboard)/dashboard/default.tsx
