+++ app/(kiosk dashboard)/dashboard/globals.css
