+++ app/(kiosk dashboard)/dashboard/header.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Toolbar from '../../components/toolbar';
+import Toolbar from '../../../components/toolbar';
 import styles from './header.module.css';
 
 export default function Header() {
