+++ app/(kiosk dashboard)/dashboard/issues/[id]/AdsPreview.module.css
