+++ app/(kiosk dashboard)/dashboard/issues/[id]/AdsPreview.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import Link from 'next/link';
-import { fetchKioskAdsByKioskId, fetchKioskById } from '../../../../helpers/httpMethods';
+import { fetchKioskAdsByKioskId, fetchKioskById } from '../../../../../helpers/httpMethods';
 import styles from './AdsPreview.module.css';
 
 interface AdsPageProps {
