+++ app/(kiosk dashboard)/dashboard/issues/[id]/InfoContainer.module.css
