+++ app/(kiosk dashboard)/dashboard/issues/[id]/Issue.module.css
