+++ app/(kiosk dashboard)/dashboard/issues/[id]/issuesList.module.css
