+++ app/(kiosk dashboard)/dashboard/issues/[id]/page.module.css
