+++ app/(kiosk dashboard)/dashboard/issues/page.tsx
