+++ app/(kiosk dashboard)/dashboard/layout.module.css
