+++ app/(kiosk dashboard)/dashboard/led/ledPreviewPlaceholder.tsx
