+++ app/(kiosk dashboard)/dashboard/page.module.css
