+++ app/(kiosk dashboard)/dashboard/page.tsx
