+++ app/(kiosk dashboard)/dashboard/sessionWrapper.tsx
