+++ app/layout.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react';
+
+export default async function RootLayout({
+	children
+}: Readonly<{
+	children: ReactNode;
+}>) {
+	return (
+		<html lang="en">
+			<body>{children}</body>
+		</html>
+	);
+}
